requirements.txt
Pillow==10.2.0
cryptography==42.0.2
numpy==1.26.4
